feat: add option to avoid user interaction

+++ scripts/10_homebrew.sh
@@ -1,10 +1,10 @@
 if ! command -v brew &> /dev/null; then
-  read -p "Press ENTER to install homebrew" < /dev/tty
+  [[ -n $FORCE ]] || read -p "Press ENTER to install homebrew" < /dev/tty
   /bin/bash -c "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/HEAD/install.sh)"
 
   eval "$(/opt/homebrew/bin/brew shellenv)"
 fi
 
-read -p "Press ENTER to brew" < /dev/tty
+[[ -n $FORCE ]] || read -p "Press ENTER to brew" < /dev/tty
 brew analytics off
 brew bundle --file=$DATA_PATH/Brewfile
